Implement 'check' target
Now it is possible to define a test to be invoked once a target has been
remade. This test can be a existing script or binary, or a target of
another module file (which then will be built accordingly).

The only (?) nasty thing in the makefile is when we try to determine if
the test is built by this project or is a preexisting program.