Modularize tests
Idea is that one or more (if connected) testcase resides in one
file, which can be executed individually or as a suite (through
run-all script).

Have one setup file which is protected by a include guard, meaning
that it will either be called once per suite or once per test case,
not both.

It requires some boiler code in each file (finding out the current
directory), other than that it works.