Add 'dist' target
The way this target is implemented is probably going to cause problems,
have not figured out a reliable way to determine *what* could possibly
go into a distribution archive without relying on a VCS to give us an
answer. Right now we search through SRCDIR and exclude things we know
that we have generated (binaries, libraries, coverage notes etc).