Fix 'check' target
When I introduced an implicit rule for linking binaries and libraries I broke
the 'check' target for modules declaring a test that does not need to be built.

This cause the script/pre-built binary to match the implicit rule, avoid this
by not listing the test as target (it will still be listed as target if it
needs to be built).