Add support for referencing project wide shared targets
Add 'link_with' keyword that matches a project unique target defined with the new
'lib' keyword. This will add basic include and linking directives to the target.

Add lib, bin and link_with keywords to handle project wide linking

This needs to be cleaned up, the combinatoric is too complex at the moment.