Update logic related to link_with
From now on, cflags/cxxflags/ldflags defined by a lib module are propagated to any
target refering to the lib will inherit these. In effect, these are part of the
public interface of the lib. Any flags that should not be propagated to dependees
should use private_cflags, private_cxxflags and private_ldflags.