Solve link_with bug
Target specific variables are propagated to pre-requisits implicitly, this caused link
directives ment for the final binary to be propagated back to the shared object causing
a build failure. Now we declare flags coming from the link_with target as private to avoid
them propagating back to where it came from in the first place.